[Ids] ahora son serverside y constantes cambiadas
