+++ src/api/brands.server.js
@@ -1,0 +1,7 @@
+import { getServerEntityById } from "./server";
+
+export const getBrandById = (id) => getServerEntityById({
+  id,
+  path: "brands",
+  responseEntity: "brand",
+});
