+++ src/api/cashBalances.server.js
@@ -1,0 +1,7 @@
+import { getServerEntityById } from "./server";
+
+export const getCashBalanceById = (id) => getServerEntityById({
+  id,
+  path: "cash-balances",
+  responseEntity: "cashBalance",
+});
