+++ src/api/customers.server.js
@@ -1,0 +1,7 @@
+import { getServerEntityById } from "./server";
+
+export const getCustomerById = (id) => getServerEntityById({
+  id,
+  path: "customers",
+  responseEntity: "customer",
+});
