+++ src/api/expenses.server.js
@@ -1,0 +1,7 @@
+import { getServerEntityById } from "./server";
+
+export const getExpenseById = (id) => getServerEntityById({
+  id,
+  path: "expenses",
+  responseEntity: "expense",
+});
