+++ src/api/products.server.js
@@ -1,0 +1,7 @@
+import { getServerEntityById } from "./server";
+
+export const getProductById = (id) => getServerEntityById({
+  id,
+  path: "products",
+  responseEntity: "product",
+});
