+++ src/api/suppliers.server.js
@@ -1,0 +1,7 @@
+import { getServerEntityById } from "./server";
+
+export const getSupplierById = (id) => getServerEntityById({
+  id,
+  path: "suppliers",
+  responseEntity: "supplier",
+});
