+++ src/app/clientes/[id]/page.jsx
@@ -1,270 +1,20 @@
-"use client";
-import { useDeleteCustomer, useEditCustomer, useGetCustomer, useSetCustomerState } from "@/api/customers";
-import { FieldsContainer, FormField, Message, MessageHeader } from "@/common/components/custom";
-import { TextField } from "@/common/components/form";
-import ModalAction from "@/common/components/modals/ModalAction";
-import UnsavedChangesModal from "@/common/components/modals/ModalUnsavedChanges";
-import { ACTIVE, COLORS, DELETE, ICONS, INACTIVE, PAGES, PLACEHOLDERS } from "@/common/constants";
-import { isItemInactive } from "@/common/utils";
-import CustomerForm from "@/components/customers/CustomerForm";
-import { Loader, useBreadcrumContext, useNavActionsContext } from "@/components/layout";
-import { useAllowUpdate, useProtectedAction, useUnsavedChanges } from "@/hooks";
-import { useMutation } from "@tanstack/react-query";
-import { useRouter } from "next/navigation";
-import { useEffect, useMemo, useRef, useState } from "react";
-import { toast } from "react-hot-toast";
+import { getCustomerById } from "@/api/customers.server";
+import { PAGE_CONSTANTS } from "@/common/constants/pages";
+import { redirect } from "next/navigation";
+import CustomerPage from "./CustomerPage";
 
-const Customer = ({ params }) => {
-  const { push } = useRouter();
-  const id = params?.id;
+const Customer = async ({ params }) => {
+  const { customer, status } = await getCustomerById(params?.id);
 
-  const { data: customer, isLoading, refetch } = useGetCustomer(id, {
-    enabled: !!id,
-  });
-  const { setLabels } = useBreadcrumContext();
-  const { resetActions, setActions } = useNavActionsContext();
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [modalAction, setModalAction] = useState(null);
-  const [activeAction, setActiveAction] = useState(null);
-  const [reason, setReason] = useState("");
-  const editCustomer = useEditCustomer();
-  const deleteCustomer = useDeleteCustomer();
-  const setCustomerState = useSetCustomerState();
-  const reasonInputRef = useRef(null);
-  const formRef = useRef(null);
-
-  const {
-    showModal: showUnsavedModal,
-    handleDiscard,
-    handleContinue,
-    onBeforeView,
-    closeModal,
-  } = useUnsavedChanges({
-    formRef,
-    onDiscard: async () => {
-      formRef.current?.resetForm();
-      setIsUpdating(false);
-    },
-  });
-
-  const { isUpdating, toggleButton, setIsUpdating } = useAllowUpdate({
-    canUpdate: true,
-    onBeforeView,
-  });
-  const { handleProtectedAction } = useProtectedAction({ formRef, onBeforeView });
-
-  useEffect(() => {
-    resetActions();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  useEffect(() => {
-    setLabels([{ name: PAGES.CUSTOMERS.NAME }, { name: customer?.name }]);
-    refetch();
-  }, [customer, setLabels, refetch]);
-
-  const modalConfig = useMemo(() => ({
-    delete: {
-      header: (
-        <>¿Está seguro que desea eliminar el cliente <i>{customer?.name} ({customer?.id}) </i> ?</>
-      ),
-      confirmText: "eliminar",
-      icon: ICONS.TRASH
-    },
-    active: {
-      header: (
-        <>¿Está seguro que desea activar el cliente <i>{customer?.name} ({customer?.id}) </i> ?</>
-      ),
-      icon: ICONS.PLAY_CIRCLE
-    },
-    inactive: {
-      header: (
-        <>¿Está seguro que desea desactivar el cliente <i>{customer?.name} ({customer?.id}) </i> ?</>
-      ),
-      icon: ICONS.PAUSE_CIRCLE
-    },
-  }), [customer]);
-
-  const handleModalClose = () => {
-    setIsModalOpen(false);
-    setModalAction(null);
-    setReason("");
-  };
-
-  const { mutate: mutateEdit, isPending: isEditPending } = useMutation({
-    mutationFn: editCustomer,
-    onSuccess: (response) => {
-      if (response.statusOk) {
-        toast.success("Cliente actualizado!");
-        setIsUpdating(false);
-      } else {
-        toast.error(`${response?.message} (${response?.error?.message})`);
-      }
-    },
-    onSettled: () => {
-      setActiveAction(null);
-      handleModalClose();
-      closeModal();
-    },
-  });
-
-  const { mutate: mutateState, isPending: isMutateStatePending } = useMutation({
-    mutationFn: setCustomerState,
-    onSuccess: (response, variables) => {
-      if (response.statusOk) {
-        toast.success(
-          variables.state === ACTIVE
-            ? 'Cliente activado!'
-            : 'Cliente desactivado!'
-        );
-      } else {
-        toast.error(`${response?.message} (${response?.error?.message})`);
-      }
-    },
-    onSettled: () => {
-      setActiveAction(null);
-      handleModalClose();
-    },
-  });
-
-  const { mutate: mutateDelete, isPending: isDeletePending } = useMutation({
-    mutationFn: () => deleteCustomer(params.id),
-    onSuccess: (response) => {
-      if (response.statusOk) {
-        toast.success("Cliente eliminado permanentemente!");
-        push(PAGES.CUSTOMERS.BASE);
-      } else {
-        toast.error(`${response?.message} (${response?.error?.message})`);
-      }
-    },
-    onSettled: () => {
-      setActiveAction(null);
-      handleModalClose();
-    },
-  });
-
-  const handleActionConfirm = async () => {
-    if (modalAction === INACTIVE && !reason) {
-      toast.error("Debe proporcionar una razón para desactivar al cliente.");
-      return;
-    }
-
-    setActiveAction(modalAction);
-
-    if (modalAction === DELETE) {
-      mutateDelete();
-    }
-
-    if (modalAction === ACTIVE || modalAction === INACTIVE) {
-      mutateState({
-        id: customer.id,
-        state: modalAction,
-        ...(modalAction === INACTIVE && { inactiveReason: reason }),
-      });
-    }
-
-    handleModalClose();
-  };
-
-  const { header, confirmText = "", icon = ICONS.QUESTION } = modalConfig[modalAction] || {};
-  const requiresConfirmation = modalAction === DELETE;
-
-  useEffect(() => {
-    const handleClick = (action) => () => handleProtectedAction(() => {
-      setModalAction(action);
-      setIsModalOpen(true);
-    });
-
-    if (customer) {
-      const actions = [
-        {
-          id: 1,
-          icon: isItemInactive(customer.state) ? ICONS.PLAY_CIRCLE : ICONS.PAUSE_CIRCLE,
-          color: COLORS.GREY,
-          onClick: handleClick(isItemInactive(customer.state) ? ACTIVE : INACTIVE),
-          text: isItemInactive(customer.state) ? "Activar" : "Desactivar",
-          loading: (activeAction === ACTIVE || activeAction === INACTIVE),
-          disabled: !!activeAction || isEditPending,
-          iconOnly: true
-        },
-        {
-          id: 2,
-          icon: ICONS.TRASH,
-          color: COLORS.RED,
-          onClick: handleClick(DELETE),
-          text: "Eliminar",
-          tooltip: customer.hasBudgets ? "No se puede eliminar este cliente, existen presupuestos asociados." : false,
-          basic: true,
-          loading: activeAction === DELETE,
-          disabled: customer.hasBudgets || !!activeAction || isEditPending,
-          iconOnly: true,
-          popupPosition: "bottom left"
-        },
-      ];
-
-      setActions(actions);
-    }
-
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [customer, activeAction, isMutateStatePending, isDeletePending, isEditPending, setActions]);
-
-  if (!isLoading && !customer) {
-    push(PAGES.NOT_FOUND.BASE);
+  if ([401, 403].includes(status)) {
+    redirect(PAGE_CONSTANTS.LOGIN.BASE);
   }
 
-  return (
-    <Loader active={isLoading || !customer}>
-      {!isItemInactive(customer?.state) && toggleButton}
-      {isItemInactive(customer?.state) && (
-        <FieldsContainer>
-          <FormField flex="1">
-            <Message negative>
-              <MessageHeader>Motivo de inactivación</MessageHeader>
-              <p>{customer.inactiveReason}</p>
-            </Message>
-          </FormField>
-          <FormField flex="1" />
-          <FormField flex="1" />
-        </FieldsContainer>
-      )}
-      <CustomerForm
-        ref={formRef}
-        customer={customer}
-        onSubmit={mutateEdit}
-        isLoading={isEditPending}
-        isUpdating={isUpdating && !isItemInactive(customer?.state)}
-        view
-        isDeletePending={isDeletePending}
-      />
-      <UnsavedChangesModal
-        open={showUnsavedModal}
-        onDiscard={handleDiscard}
-        onContinue={handleContinue}
-      />
-      <ModalAction
-        title={header}
-        onConfirm={handleActionConfirm}
-        confirmationWord={requiresConfirmation ? confirmText : ""}
-        confirmButtonIcon={icon}
-        showModal={isModalOpen}
-        setShowModal={handleModalClose}
-        isLoading={isDeletePending || isMutateStatePending}
-        noConfirmation={!requiresConfirmation}
-        disableButtons={!reason && modalAction === INACTIVE}
-        reasonInputRef={reasonInputRef}
-        bodyContent={
-          modalAction === INACTIVE && (
-            <TextField
-              ref={reasonInputRef}
-              placeholder={PLACEHOLDERS.REASON}
-              value={reason}
-              onChange={(e) => setReason(e.target.value)}
-            />
-          )
-        }
-      />
-    </Loader>
-  );
+  if (!customer) {
+    redirect(PAGE_CONSTANTS.NOT_FOUND.BASE);
+  }
+
+  return <CustomerPage customer={customer} />;
 };
 
 export default Customer;
