+++ src/app/page.tsx
@@ -1,5 +1,6 @@
+import { PAGE_CONSTANTS } from "@/common/constants/pages";
 import { redirect } from "next/navigation";
 
 export default function Home() {
-  redirect("/ventas");
+  redirect(PAGE_CONSTANTS.BUDGETS.BASE);
 }
