+++ src/app/proveedores/[id]/page.jsx
@@ -1,492 +1,20 @@
-"use client";
-import { useUserContext } from "@/User";
-import { useDeleteBySupplierId, useProductsBySupplierId } from "@/api/products";
-import { useDeleteSupplier, useEditSupplier, useGetSupplier, useSetSupplierState } from "@/api/suppliers";
-import { Button, DropdownItem, FieldsContainer, FormField, Icon, Message, MessageHeader, } from "@/common/components/custom";
-import PrintBarCodes from "@/common/components/custom/PrintBarCodes";
-import { TextField } from "@/common/components/form";
-import { ModalAction } from "@/common/components/modals";
-import UnsavedChangesModal from "@/common/components/modals/ModalUnsavedChanges";
-import { ACTIVE, COLORS, ICONS, INACTIVE, PAGES, PLACEHOLDERS, SIZES } from "@/common/constants";
-import { downloadExcel, isItemInactive } from "@/common/utils";
-import { Loader, OnlyPrint, useBreadcrumContext, useNavActionsContext, } from "@/components/layout";
-import { BatchImportStock } from "@/components/products/BatchImportStock";
-import { DISCOUNT_STOCK, PRODUCT_STATES, UPLOAD_STOCK } from "@/components/products/products.constants";
-import { getFormatedMargin } from "@/components/products/products.utils";
-import SupplierForm from "@/components/suppliers/SupplierForm";
-import { EXAMPLE_TEMPLATE_DATA_STOCK } from "@/components/suppliers/suppliers.constants";
-import { useAllowUpdate, useProtectedAction, useUnsavedChanges } from "@/hooks";
-import { RULES } from "@/roles";
-import { useMutation } from "@tanstack/react-query";
-import { useRouter } from "next/navigation";
-import { useEffect, useMemo, useRef, useState } from "react";
-import { toast } from "react-hot-toast";
-import { useReactToPrint } from "react-to-print";
-import { Dropdown, Popup } from "semantic-ui-react";
+import { getSupplierById } from "@/api/suppliers.server";
+import { PAGE_CONSTANTS } from "@/common/constants/pages";
+import { redirect } from "next/navigation";
+import SupplierPage from "./SupplierPage";
 
-const Supplier = ({ params }) => {
-  const { role } = useUserContext();
-  const { push } = useRouter();
-  const { data: supplier, isLoading, refetch: refetchSupplier } = useGetSupplier(params.id);
-  const { data: products, isLoading: loadingProducts, refetch: refetchProducts } = useProductsBySupplierId(params.id);
-  const { setLabels } = useBreadcrumContext();
-  const { resetActions, setActions } = useNavActionsContext();
-  const [confirmationAction, setConfirmationAction] = useState(null);
-  const [activeAction, setActiveAction] = useState(null);
-  const [isExcelLoading, setIsExcelLoading] = useState(false);
-  const [reason, setReason] = useState("");
-  const printRef = useRef(null);
-  const formRef = useRef(null);
-  const {
-    showModal: showUnsavedModal,
-    handleDiscard,
-    handleContinue,
-    onBeforeView,
-  } = useUnsavedChanges({
-    formRef,
-    onDiscard: async () => {
-      formRef.current?.resetForm();
-      setIsUpdating(false);
-    },
-  });
-  const { isUpdating, toggleButton, setIsUpdating } = useAllowUpdate({
-    canUpdate: RULES.canUpdate[role],
-    onBeforeView,
-  });
-  const { handleProtectedAction } = useProtectedAction({ formRef, onBeforeView });
-  const editSupplier = useEditSupplier();
-  const deleteSupplier = useDeleteSupplier();
-  const deleteBySupplierId = useDeleteBySupplierId();
-  const setSupplierState = useSetSupplierState();
+const Supplier = async ({ params }) => {
+  const { supplier, status } = await getSupplierById(params?.id);
 
-  useEffect(() => {
-    resetActions();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  useEffect(() => {
-    setLabels([{ name: PAGES.SUPPLIERS.NAME }, { name: supplier?.name }]);
-    refetchSupplier();
-    refetchProducts();
-  }, [setLabels, supplier, refetchProducts, refetchSupplier]);
-
-  const hasAssociatedProducts = useMemo(() => !!products?.length, [products]);
-
-  const handlePrint = useReactToPrint({
-    content: () => printRef.current,
-    removeAfterPrint: true,
-  });
-
-  const modalConfig = {
-    deleteSupplier: {
-      header: (
-        <>¿Está seguro que desea eliminar PERMANENTEMENTE al proveedor <i>{supplier?.name} ({supplier?.id}) </i> ?</>
-      ),
-      confirmText: "eliminar",
-      icon: ICONS.TRASH,
-      tooltip: hasAssociatedProducts
-        ? 'No se puede eliminar este proveedor, existen productos asociados.'
-        : false,
-    },
-    deleteBatch: {
-      header: (
-        <>¿Está seguro que desea eliminar los {products?.length || ""} productos del proveedor <i>{supplier?.name} ({supplier?.id}) </i> ?</>
-      ),
-      confirmText: "eliminar",
-      icon: ICONS.TRASH,
-    },
-    active: {
-      header: (
-        <>¿Está seguro que desea activar el proveedor <i>{supplier?.name} ({supplier?.id}) </i> ?</>
-      ),
-      icon: ICONS.PLAY_CIRCLE,
-    },
-    inactive: {
-      header: (
-        <>¿Está seguro que desea desactivar el proveedor <i>{supplier?.name} ({supplier?.id}) </i> ?</>
-      ),
-      icon: ICONS.PAUSE_CIRCLE,
-    },
-  };
-
-  const handleModalClose = () => {
-    setConfirmationAction(null);
-    setReason("");
-  };
-
-  const { mutate: mutateEdit, isPending: isEditPending } = useMutation({
-    mutationFn: editSupplier,
-    onSuccess: (response) => {
-      if (response.statusOk) {
-        toast.success('Proveedor actualizado!');
-        setIsUpdating(false);
-      } else {
-        toast.error(`${response?.message} (${response?.error?.message})`);
-      }
-    },
-    onSettled: () => {
-      setActiveAction(null);
-      handleModalClose();
-      closeModal();
-    },
-  });
-
-  const { mutate: mutateDeleteBatch, isPending: isDeleteBatchPending } =
-    useMutation({
-      mutationFn: () => deleteBySupplierId(params.id),
-      onSuccess: (response) => {
-        if (response.statusOk) {
-          toast.success('Lista de productos del proveedor eliminada!');
-          handleModalClose();
-        } else {
-          toast.error(`${response?.message} (${response?.error?.message})`);
-        }
-      },
-      onSettled: () => {
-        setActiveAction(null);
-        handleModalClose();
-      },
-    });
-
-  const { mutate: mutateDelete, isPending: isDeletePending } =
-    useMutation({
-      mutationFn: () => deleteSupplier(params.id),
-      onSuccess: (response) => {
-        if (response.statusOk) {
-          toast.success('Proveedor eliminado permanentemente!');
-          push(PAGES.SUPPLIERS.BASE);
-        } else {
-          toast.error(`${response?.message} (${response?.error?.message})`);
-        }
-      },
-      onSettled: () => {
-        setActiveAction(null);
-        handleModalClose();
-      },
-    });
-
-  const { mutate: mutateState, isPending: isMutateStatePending } = useMutation({
-    mutationFn: setSupplierState,
-    onSuccess: (response, variables) => {
-      if (response.statusOk) {
-        toast.success(
-          variables.state === ACTIVE
-            ? 'Proveedor activado!'
-            : 'Proveedor desactivado!'
-        );
-      } else {
-        toast.error(`${response?.message} (${response?.error?.message})`);
-      }
-    },
-    onSettled: () => {
-      setActiveAction(null);
-      handleModalClose();
-    },
-  });
-
-  const handleActionConfirm = async () => {
-    if (confirmationAction === INACTIVE && !reason) {
-      toast.error("Debe proporcionar una razón para desactivar el proveedor.");
-      return;
-    }
-
-    setActiveAction(confirmationAction);
-
-    if (confirmationAction === "deleteBatch") {
-      mutateDeleteBatch();
-    }
-
-    if (confirmationAction === "deleteSupplier") {
-      mutateDelete();
-    }
-
-    if (confirmationAction === ACTIVE || confirmationAction === INACTIVE) {
-      mutateState({
-        id: supplier.id,
-        state: confirmationAction,
-        ...(confirmationAction === INACTIVE && { inactiveReason: reason }),
-      });
-    }
-
-    handleModalClose();
-  };
-
-  const { header, confirmText = '', icon = ICONS.QUESTION } =
-    modalConfig[confirmationAction] || {};
-
-  const requiresConfirmation =
-    confirmationAction === "deleteSupplier" ||
-    confirmationAction === "deleteBatch";
-
-  useEffect(() => {
-    const handleBarCodePrint = () => {
-      if (hasAssociatedProducts) {
-        setActiveAction("print");
-        handlePrint();
-        setActiveAction(null);
-      } else {
-        toast("No hay productos de este proveedor.", {
-          icon: (
-            <Icon
-              margin="0"
-              toast
-              name={ICONS.INFO_CIRCLE}
-              color={COLORS.BLUE}
-            />
-          ),
-        });
-      }
-    };
-
-    const handleDownloadExcel = () => {
-      if (!products) return;
-      const headers = [
-        'Id',
-        'Nombre',
-        'Marca',
-        'Proveedor',
-        'Costo',
-        'Precio',
-        'Margen',
-        'Estado',
-        'Comentarios',
-      ];
-
-      const mappedProducts = products.map((product) => {
-        const productState = PRODUCT_STATES[product.state]?.singularTitle || product.state;
-        return [
-          product.id,
-          product.name,
-          product.brandName,
-          product.supplierName,
-          product.cost,
-          product.price,
-          getFormatedMargin(product.price, product.cost),
-          productState,
-          product.comments,
-        ];
-      });
-      downloadExcel(
-        [headers, ...mappedProducts],
-        `Lista de productos de ${supplier.name}`,
-      );
-    };
-
-    let actions = [];
-
-    if (RULES.canRemove[role]) {
-      const handleClick = (action) => () =>
-        handleProtectedAction(() => {
-          setConfirmationAction(action);
-        });
-      const isStockExcelDisabled =
-        isExcelLoading ||
-        !!activeAction ||
-        loadingProducts ||
-        isEditPending ||
-        !hasAssociatedProducts;
-      const stockExcelTooltip = !loadingProducts && !hasAssociatedProducts
-        ? "No se puede importar stock porque no existen productos de este proveedor."
-        : false;
-
-      actions = [{
-        id: 1,
-        icon: ICONS.BARCODE,
-        color: COLORS.BLUE,
-        text: "Imprimir códigos de barra",
-        onClick: handleBarCodePrint,
-        loading: activeAction === "print",
-        disabled: !!activeAction || isEditPending || !hasAssociatedProducts,
-        tooltip: !hasAssociatedProducts
-          ? 'No existen productos de este proveedor.'
-          : false,
-        iconOnly: true
-      },
-      {
-        id: 2,
-        icon: isItemInactive(supplier?.state)
-          ? ICONS.PLAY_CIRCLE
-          : ICONS.PAUSE_CIRCLE,
-        color: COLORS.GREY,
-        text: isItemInactive(supplier?.state) ? "Activar" : "Desactivar",
-        onClick: handleClick(isItemInactive(supplier?.state) ? ACTIVE : INACTIVE),
-        loading: activeAction === ACTIVE || activeAction === INACTIVE,
-        disabled: !!activeAction || isEditPending,
-        iconOnly: true
-      },
-      {
-        id: 3,
-        iconOnly: true,
-        button: (
-          <Popup
-            disabled={!stockExcelTooltip}
-            position="top center"
-            size={SIZES.TINY}
-            content={(
-              <div>
-                <div><strong>Stock/Excel</strong></div>
-                <div>{stockExcelTooltip}</div>
-              </div>
-            )}
-            trigger={(
-              <div>
-                <Dropdown
-                  pointing
-                  as={Button}
-                  text="Stock/Excel"
-                  icon={ICONS.FILE_EXCEL}
-                  width="fit-content"
-                  floating
-                  labeled
-                  button
-                  className="icon"
-                  disabled={isStockExcelDisabled}
-                >
-                  <Dropdown.Menu>
-                    <DropdownItem
-                      onClick={() => {
-                        if (products?.length) {
-                          setIsExcelLoading(true);
-                          handleDownloadExcel();
-                          setIsExcelLoading(false);
-                        } else {
-                          toast("No hay productos de este proveedor para descargar.", {
-                            icon: (
-                              <Icon
-                                margin="0"
-                                toast
-                                name={ICONS.INFO_CIRCLE}
-                                color={COLORS.BLUE}
-                              />
-                            ),
-                          });
-                        }
-                      }}
-                    >
-                      <Icon name={ICONS.DOWNLOAD} color="blue" />
-                      Descargar productos
-                    </DropdownItem>
-                    <DropdownItem  >
-                      <BatchImportStock
-                        mode={UPLOAD_STOCK}
-                        supplierId={supplier?.id}
-                        products={products}
-                      />
-                    </DropdownItem>
-                    <DropdownItem  >
-                      <BatchImportStock
-                        mode={DISCOUNT_STOCK}
-                        supplierId={supplier?.id}
-                        products={products}
-                      />
-                    </DropdownItem>
-                    <DropdownItem onClick={() => downloadExcel(EXAMPLE_TEMPLATE_DATA_STOCK, "Ejemplo de tabla stock")}>
-                      <Icon name={ICONS.FILE_EXCEL_OUTLINE} />Plantilla stock
-                    </DropdownItem>
-                  </Dropdown.Menu>
-                </Dropdown>
-              </div>
-            )}
-          />
-        ),
-      },
-      {
-        id: 4,
-        icon: ICONS.LIST_UL,
-        color: COLORS.RED,
-        text: "Eliminar todos los productos del proveedor",
-        onClick: handleClick('deleteBatch'),
-        loading: activeAction === "deleteBatch",
-        disabled: !hasAssociatedProducts || !!activeAction || isEditPending,
-        tooltip: !hasAssociatedProducts
-          ? 'No existen productos de este proveedor.'
-          : false,
-        iconOnly: true
-      },
-      {
-        id: 5,
-        icon: ICONS.TRASH,
-        color: COLORS.RED,
-        text: "Eliminar proveedor",
-        onClick: handleClick('deleteSupplier'),
-        loading: activeAction === "deleteSupplier",
-        disabled: hasAssociatedProducts || !!activeAction || isEditPending,
-        tooltip: hasAssociatedProducts
-          ? 'No se puede eliminar este proveedor, existen productos asociados.'
-          : false,
-        iconOnly: true,
-        basic: true,
-        popupPosition: "bottom right"
-      }];
-    }
-
-    setActions(actions);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [role, supplier, products, hasAssociatedProducts, activeAction, isEditPending, loadingProducts, isExcelLoading]);
-
-  if (!isLoading && !supplier) {
-    push(PAGES.NOT_FOUND.BASE);
+  if ([401, 403].includes(status)) {
+    redirect(PAGE_CONSTANTS.LOGIN.BASE);
   }
 
-  return (
-    <Loader active={isLoading || loadingProducts || !supplier}>
-      {!isItemInactive(supplier?.state) && toggleButton}
-      {isItemInactive(supplier?.state) && (
-        <FieldsContainer>
-          <FormField flex="1">
-            <Message negative>
-              <MessageHeader>Motivo de inactivación</MessageHeader>
-              <p>{supplier.inactiveReason}</p>
-            </Message>
-          </FormField>
-          <FormField flex="1" />
-          <FormField flex="1" />
-        </FieldsContainer>
-      )}
-      <SupplierForm
-        ref={formRef}
-        supplier={supplier}
-        onSubmit={mutateEdit}
-        isLoading={isEditPending}
-        isUpdating={isUpdating && !isItemInactive(supplier?.state)}
-        view
-        isDeletePending={isDeletePending}
-      />
-      {!isUpdating && (
-        <OnlyPrint>
-          <PrintBarCodes ref={printRef} products={products} />
-        </OnlyPrint>
-      )}
-      <UnsavedChangesModal
-        open={showUnsavedModal}
-        onDiscard={handleDiscard}
-        onContinue={handleContinue}
-      />
-      <ModalAction
-        title={header}
-        onConfirm={handleActionConfirm}
-        confirmationWord={requiresConfirmation ? confirmText : ""}
-        confirmButtonIcon={icon}
-        showModal={!!confirmationAction}
-        setShowModal={handleModalClose}
-        isLoading={
-          isDeleteBatchPending ||
-          isDeletePending ||
-          isMutateStatePending
-        }
-        noConfirmation={!requiresConfirmation}
-        bodyContent={
-          confirmationAction === INACTIVE && (
-            <TextField
-              placeholder={PLACEHOLDERS.REASON}
-              value={reason}
-              onChange={(e) => setReason(e.target.value)}
-            />
-          )
-        }
-      />
-    </Loader>
-  );
+  if (!supplier) {
+    redirect(PAGE_CONSTANTS.NOT_FOUND.BASE);
+  }
+
+  return <SupplierPage supplier={supplier} />;
 };
 
 export default Supplier;
