+++ src/common/constants/pages.js
@@ -1,0 +1,84 @@
+export const PAGE_CONSTANTS = {
+  BASE: "/",
+  CUSTOMERS: {
+    BASE: "/clientes",
+    CREATE: "/clientes/crear",
+    NAME: "Clientes",
+    SINGULAR_NAME: "Cliente",
+    SHORTKEYS: "Control+1",
+  },
+  SUPPLIERS: {
+    BASE: "/proveedores",
+    CREATE: "/proveedores/crear",
+    NAME: "Proveedores",
+    SINGULAR_NAME: "Proveedor",
+    SHORTKEYS: "Control+2",
+  },
+  BRANDS: {
+    BASE: "/marcas",
+    CREATE: "/marcas/crear",
+    NAME: "Marcas",
+    SINGULAR_NAME: "Marca",
+    SHORTKEYS: "Control+3",
+  },
+  PRODUCTS: {
+    BASE: "/productos",
+    CREATE: "/productos/crear",
+    NAME: "Productos",
+    SINGULAR_NAME: "Producto",
+    SHORTKEYS: "Control+4",
+  },
+  BUDGETS: {
+    BASE: "/ventas",
+    CREATE: "/ventas/crear",
+    NAME: "Ventas",
+    SINGULAR_NAME: "Venta",
+    SHORTKEYS: "Control+5",
+    DRAFT_SEGMENT: "borrador",
+  },
+  EXPENSES: {
+    BASE: "/gastos",
+    CREATE: "/gastos/crear",
+    NAME: "Gastos",
+    SINGULAR_NAME: "Gasto",
+    SHORTKEYS: "Control+6",
+  },
+  CASH_BALANCES: {
+    BASE: "/cajas",
+    CREATE: "/cajas/crear",
+    NAME: "Cajas",
+    SINGULAR_NAME: "Caja",
+    SHORTKEYS: "Control+7",
+  },
+  USERS: {
+    BASE: "/usuarios",
+    CREATE: "/usuarios/crear",
+    NAME: "Usuarios",
+    SINGULAR_NAME: "Usuario",
+    SHORTKEYS: "Control+8",
+  },
+  BUDGETS_HISTORY: {
+    BASE: "/historial-ventas",
+  },
+  LOGIN: {
+    BASE: "/login",
+  },
+  SETTINGS: {
+    BASE: "/configuracion",
+    NAME: "Configuración",
+    SHORTKEYS: "Control+9",
+  },
+  CHANGE_PASSWORD: {
+    BASE: "/cambiar-contrasena",
+  },
+  RESTORE_PASSWORD: {
+    BASE: "/recuperar-contrasena",
+  },
+  MAINTENANCE: {
+    BASE: "/mantenimiento",
+    NAME: "Mantenimiento",
+  },
+  NOT_FOUND: {
+    BASE: "/ups",
+  },
+};
